backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pdfplumber==0.11.0
python-docx==1.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
sqlalchemy==2.0.35
aiofiles==24.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
